# docs/swagger.yaml
basePath: /api
definitions:
...
  controller.ResponseData:
    properties:
      code:
        type: string
      data: {}
      msg:
        type: string
    type: object
...
  controller.reqCommentInfo:
    properties:
      ParentCommentId:
        type: integer
      articleId:
        type: integer
      avatar:
        type: string
      content:
        type: string
      email:
        type: string
      nickname:
        type: string
      replyCommentId:
        type: integer
      website:
        type: string
    required:
    - articleId
    - content
    - nickname
    type: object
...
  dto.CommentDto:
    properties:
      avatar:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      nickname:
        type: string
      replyNickname:
        description: TODO show reply nickname
        type: string
      subComments:
        items:
          $ref: '#/definitions/dto.CommentDto'
        type: array
      website:
        type: string
    type: object
  dto.CommentTreeDto:
    properties:
      comments:
        items:
          $ref: '#/definitions/dto.CommentDto'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: victorzhoux@163.com
    name: VictorZhou
  title: vicBlog server API
  version: "1.0"
paths:
...
  /v1/comment:
    post:
      consumes:
      - application/json
      description: add a comment
      parameters:
      - description: body of add comment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.reqCommentInfo'
      responses:
        "201":
          description: Created
      summary: add comment
      tags:
      - Comment
  /v1/comment/:articleId:
    get:
      consumes:
      - application/json
      description: get all comments and sort them as tree like format
      parameters:
      - description: router url
        in: path
        name: articleId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommentTreeDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: Get Comments Tree
      tags:
      - Comment
...
swagger: "2.0"